requirements.txt
fastapi
uvicorn
langchain
faiss-cpu
sentence-transformers
PyMuPDF
langdetect
scikit-learn
